pyproject.toml
[tool.poetry]
name = "opendata"
version = "0.1.0"
description = "An open data platform with a gamification twist."
authors = [
    "Martin Balfroid <martin.balfroid@student.unamur.be>",
    "Maxime Caucheteur <maxime.caucheteur@student.unamur.be>",
    "Jonathan Ferreira <jonathan.ferreira@student.unamur.be>",
    "Germain Herbay <germain.herbay@student.unamur.be>",
    "Loris Hincq <loris.hincq@student.unamur.be>",
    "Pierre Luycx <pierre.luycx@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-colorfield = "^0.4.1"
django-nested-inline = "^0.4.2"
django-notifications-hq = "^1.6.0"
requests = "^2.25.1"
dictor = "^0.1.7"
Pillow = "^8.2.0"
spacy = "^3.0.5"
fr-core-news-sm = {url = "https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0.tar.gz"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
